# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Matcha API","description":"API for Matcha dating application","version":"1.0.0"},"paths":{"/api/auth/signup":{"post":{"tags":["Authentication"],"summary":"Signup","operationId":"signup_api_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Invalid credentials"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/api/auth/signin":{"post":{"tags":["Authentication"],"summary":"Signin","operationId":"signin_api_auth_signin_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Invalid credentials"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/api/auth/email-verification":{"get":{"tags":["Authentication"],"summary":"Email Verification","operationId":"email_verification_api_auth_email_verification_get","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Invalid credentials"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/api/auth/forgot-password":{"post":{"tags":["Authentication"],"summary":"Forgot Password","operationId":"forgot_password_api_auth_forgot_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordResetRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Invalid credentials"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/api/auth/validate-reset-token/{token}":{"get":{"tags":["Authentication"],"summary":"Validate Reset Token","operationId":"validate_reset_token_api_auth_validate_reset_token__token__get","parameters":[{"name":"token","in":"path","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Invalid credentials"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/api/auth/reset-password/{token}":{"post":{"tags":["Authentication"],"summary":"Reset Password","operationId":"reset_password_api_auth_reset_password__token__post","parameters":[{"name":"token","in":"path","required":true,"schema":{"type":"string","title":"Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordResetConfirm"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Invalid credentials"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/api/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Me","operationId":"get_me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Invalid credentials"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/api/auth/signout":{"post":{"tags":["Authentication"],"summary":"Signout","operationId":"signout_api_auth_signout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Invalid credentials"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/api/users/suggestions":{"get":{"tags":["User Management"],"summary":"Advanced profile search","description":"Search and browse profiles matching mutual preferences. Supports filtering by age, fame, distance, and interest tags, with sorting by age, location, fame, or shared interests.","operationId":"get_suggestions_api_users_suggestions_get","parameters":[{"name":"sort_by","in":"query","required":false,"schema":{"$ref":"#/components/schemas/SortBy","default":"distance"}},{"name":"order","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/SortOrder"},{"type":"null"}],"title":"Order"}},{"name":"min_age","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":120,"minimum":18},{"type":"null"}],"title":"Min Age"}},{"name":"max_age","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":120,"minimum":18},{"type":"null"}],"title":"Max Age"}},{"name":"max_distance_km","in":"query","required":false,"schema":{"anyOf":[{"type":"number","exclusiveMinimum":0},{"type":"null"}],"title":"Max Distance Km"}},{"name":"min_fame","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","minimum":0},{"type":"null"}],"title":"Min Fame"}},{"name":"max_fame","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","minimum":0},{"type":"null"}],"title":"Max Fame"}},{"name":"tags","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tags"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuggestionListResponse"}}}},"401":{"description":"Unauthorized - Authentication required"},"403":{"description":"Forbidden - Insufficient permissions"},"404":{"description":"User not found"},"422":{"description":"Validation error"}}}},"/api/users/me/viewers":{"get":{"tags":["User Management"],"summary":"Get profile viewers","description":"List users who viewed the authenticated user's profile","operationId":"get_my_viewers_api_users_me_viewers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Authentication required"},"403":{"description":"Forbidden - Insufficient permissions"},"404":{"description":"User not found"},"422":{"description":"Validation error"}}}},"/api/users/me/likes":{"get":{"tags":["User Management"],"summary":"Get users who liked me","description":"List users who liked the authenticated user's profile","operationId":"get_my_likes_api_users_me_likes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Authentication required"},"403":{"description":"Forbidden - Insufficient permissions"},"404":{"description":"User not found"},"422":{"description":"Validation error"}}}},"/api/users/{username}":{"get":{"tags":["User Management"],"summary":"Get User Profile","description":"Retrieve public profile information for a specific user","operationId":"get_user_api_users__username__get","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}},"401":{"description":"Unauthorized - Authentication required"},"403":{"description":"Forbidden - Insufficient permissions"},"404":{"description":"User not found"},"422":{"description":"Validation error"}}},"patch":{"tags":["User Management"],"summary":"Update User Profile","description":"Update profile information for the authenticated user","operationId":"partial_update_user_api_users__username__patch","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}},"401":{"description":"Unauthorized - Authentication required"},"403":{"description":"Forbidden - Insufficient permissions"},"404":{"description":"User not found"},"422":{"description":"Validation error"}}},"put":{"tags":["User Management"],"summary":"Replace User Profile","description":"Replace profile information for the authenticated user","operationId":"update_user_api_users__username__put","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPut"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}},"401":{"description":"Unauthorized - Authentication required"},"403":{"description":"Forbidden - Insufficient permissions"},"404":{"description":"User not found"},"422":{"description":"Validation error"}}},"delete":{"tags":["User Management"],"summary":"Delete User Profile","description":"Delete the authenticated user's profile","operationId":"delete_user_api_users__username__delete","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"User profile deleted","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Authentication required"},"403":{"description":"Forbidden - Insufficient permissions"},"404":{"description":"User not found"},"422":{"description":"Validation error"}}}},"/api/users/{username}/like":{"post":{"tags":["User Management"],"summary":"Like a user","operationId":"like_user_api_users__username__like_post","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Authentication required"},"403":{"description":"Forbidden - Insufficient permissions"},"404":{"description":"User not found"},"422":{"description":"Validation error"}}},"delete":{"tags":["User Management"],"summary":"Unlike a user","operationId":"unlike_user_api_users__username__like_delete","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Authentication required"},"403":{"description":"Forbidden - Insufficient permissions"},"404":{"description":"User not found"},"422":{"description":"Validation error"}}}},"/api/users/me/connections":{"get":{"tags":["Social"],"summary":"Get My Connections","operationId":"get_my_connections_api_users_me_connections_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/users/{username}/block":{"post":{"tags":["Social"],"summary":"Block User","operationId":"block_user_api_users__username__block_post","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Social"],"summary":"Unblock User","operationId":"unblock_user_api_users__username__block_delete","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/users/{username}/report":{"post":{"tags":["Social"],"summary":"Report User","operationId":"report_user_api_users__username__report_post","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chat/{username}/messages":{"get":{"tags":["Social"],"summary":"Get Chat Messages","operationId":"get_chat_messages_api_chat__username__messages_get","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatMessageResponse"},"title":"Response Get Chat Messages Api Chat  Username  Messages Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Social"],"summary":"Send Chat Message","operationId":"send_chat_message_api_chat__username__messages_post","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessageCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/notifications":{"get":{"tags":["Notifications"],"summary":"List Notifications","operationId":"list_notifications_api_notifications_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/NotificationResponse"},"type":"array","title":"Response List Notifications Api Notifications Get"}}}}}}},"/api/notifications/unread-count":{"get":{"tags":["Notifications"],"summary":"Get Unread Count","operationId":"get_unread_count_api_notifications_unread_count_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/notifications/{notification_id}/read":{"patch":{"tags":["Notifications"],"summary":"Mark Notification Read","operationId":"mark_notification_read_api_notifications__notification_id__read_patch","parameters":[{"name":"notification_id","in":"path","required":true,"schema":{"type":"integer","title":"Notification Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/notifications/read-all":{"post":{"tags":["Notifications"],"summary":"Mark All Notifications Read","operationId":"mark_all_notifications_read_api_notifications_read_all_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/upload":{"post":{"tags":["File Upload"],"summary":"Upload Images","description":"Upload up to 5 image files with validation","operationId":"upload_files_api_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_files_api_upload_post"}}},"required":true},"responses":{"201":{"description":"Uploaded image details with URLs","content":{"application/json":{"schema":{}}}},"400":{"description":"Invalid file upload"},"401":{"description":"Unauthorized - Authentication required"},"500":{"description":"Server error during file upload"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/images/{path}":{"get":{"tags":["Images"],"summary":"Get Image","operationId":"get_image_api_images__path__get","parameters":[{"name":"path","in":"path","required":true,"schema":{"type":"string","title":"Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/interests":{"get":{"tags":["Interests"],"summary":"List Interests","operationId":"list_interests_api_interests_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/location/geocode":{"post":{"tags":["Location"],"summary":"Geocode Location","operationId":"geocode_location_api_location_geocode_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeocodeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeocodeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_files_api_upload_post":{"properties":{"files":{"items":{"type":"string","contentMediaType":"application/octet-stream"},"type":"array","title":"Files"}},"type":"object","required":["files"],"title":"Body_upload_files_api_upload_post"},"ChatMessageCreate":{"properties":{"body":{"type":"string","maxLength":2000,"minLength":1,"title":"Body"}},"type":"object","required":["body"],"title":"ChatMessageCreate"},"ChatMessageResponse":{"properties":{"id":{"type":"integer","title":"Id"},"body":{"type":"string","title":"Body"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"read_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Read At"},"is_mine":{"type":"boolean","title":"Is Mine"},"sender_username":{"type":"string","title":"Sender Username"}},"type":"object","required":["id","body","created_at","is_mine","sender_username"],"title":"ChatMessageResponse"},"Gender":{"type":"string","enum":["Male","Female"],"title":"Gender"},"GeocodeRequest":{"properties":{"query":{"type":"string","maxLength":128,"minLength":2,"title":"Query"}},"type":"object","required":["query"],"title":"GeocodeRequest"},"GeocodeResponse":{"properties":{"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"},"label":{"type":"string","title":"Label"}},"type":"object","required":["latitude","longitude","label"],"title":"GeocodeResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"NotificationActor":{"properties":{"username":{"type":"string","title":"Username"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"profile_picture":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Picture"}},"type":"object","required":["username","first_name","last_name"],"title":"NotificationActor"},"NotificationResponse":{"properties":{"id":{"type":"integer","title":"Id"},"type":{"type":"string","title":"Type"},"payload":{"additionalProperties":true,"type":"object","title":"Payload","default":{}},"read_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Read At"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"actor":{"$ref":"#/components/schemas/NotificationActor"}},"type":"object","required":["id","type","created_at","actor"],"title":"NotificationResponse"},"PasswordResetConfirm":{"properties":{"new_password":{"type":"string","title":"New Password"}},"type":"object","required":["new_password"],"title":"PasswordResetConfirm"},"PasswordResetRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["email"],"title":"PasswordResetRequest"},"ReportRequest":{"properties":{"reason":{"type":"string","maxLength":512,"minLength":5,"title":"Reason"}},"type":"object","required":["reason"],"title":"ReportRequest"},"SignInData":{"properties":{"login":{"anyOf":[{"type":"string"},{"type":"string","format":"email"}],"title":"Login"},"password":{"type":"string","minLength":8,"title":"Password"}},"type":"object","required":["login","password"],"title":"SignInData"},"SignupData":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"gender":{"$ref":"#/components/schemas/Gender"},"birthdate":{"type":"string","format":"date","title":"Birthdate"}},"type":"object","required":["username","email","password","first_name","last_name","gender","birthdate"],"title":"SignupData"},"SortBy":{"type":"string","enum":["distance","age","fame","common_tags"],"title":"SortBy"},"SortOrder":{"type":"string","enum":["asc","desc"],"title":"SortOrder"},"SuggestedProfile":{"properties":{"username":{"type":"string","title":"Username"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"age":{"type":"integer","title":"Age"},"bio":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bio"},"fame_rating":{"type":"integer","title":"Fame Rating","default":0},"profile_picture":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Picture"},"interests":{"items":{"type":"string"},"type":"array","title":"Interests","default":[]},"common_interests":{"items":{"type":"string"},"type":"array","title":"Common Interests","default":[]},"common_interest_count":{"type":"integer","title":"Common Interest Count","default":0},"distance_km":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Distance Km"},"location_label":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location Label"}},"type":"object","required":["username","first_name","last_name","age"],"title":"SuggestedProfile"},"SuggestionListResponse":{"properties":{"results":{"items":{"$ref":"#/components/schemas/SuggestedProfile"},"type":"array","title":"Results"},"total":{"type":"integer","title":"Total"},"limit":{"type":"integer","title":"Limit"},"offset":{"type":"integer","title":"Offset"}},"type":"object","required":["results","total","limit","offset"],"title":"SuggestionListResponse"},"UserImage":{"properties":{"url":{"type":"string","title":"Url"},"is_profile_picture":{"type":"boolean","title":"Is Profile Picture","default":false}},"type":"object","required":["url"],"title":"UserImage"},"UserProfileResponse":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"username":{"type":"string","title":"Username"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"gender":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Gender"},"bio":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bio"},"sexual_preference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sexual Preference"},"latitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Latitude"},"longitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Longitude"},"location_label":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location Label"},"is_verified":{"type":"boolean","title":"Is Verified"},"is_profile_completed":{"type":"boolean","title":"Is Profile Completed"},"fame_rating":{"type":"integer","title":"Fame Rating","default":0},"images":{"items":{"$ref":"#/components/schemas/UserImage"},"type":"array","title":"Images","default":[]},"interests":{"items":{"type":"string"},"type":"array","title":"Interests","default":[]},"profile_picture":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Picture"},"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age"},"is_liked_by_viewer":{"type":"boolean","title":"Is Liked By Viewer","default":false},"has_liked_viewer":{"type":"boolean","title":"Has Liked Viewer","default":false},"is_connected":{"type":"boolean","title":"Is Connected","default":false},"is_blocked":{"type":"boolean","title":"Is Blocked","default":false},"is_online":{"type":"boolean","title":"Is Online","default":false},"last_seen_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Seen At"}},"type":"object","required":["id","username","first_name","last_name","is_verified","is_profile_completed"],"title":"UserProfileResponse"},"UserPut":{"properties":{"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"email":{"type":"string","format":"email","title":"Email"},"bio":{"type":"string","maxLength":255,"minLength":10,"title":"Bio"},"gender":{"$ref":"#/components/schemas/Gender"},"sexual_preference":{"$ref":"#/components/schemas/Gender"},"interests":{"anyOf":[{"items":{"type":"string"},"type":"array","maxItems":5},{"type":"null"}],"title":"Interests"},"images":{"anyOf":[{"items":{"type":"string"},"type":"array","maxItems":5},{"type":"null"}],"title":"Images"},"profile_picture":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Picture"},"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"},"location_label":{"anyOf":[{"type":"string","maxLength":128},{"type":"null"}],"title":"Location Label"},"username":{"type":"string","title":"Username"}},"type":"object","required":["first_name","last_name","email","bio","gender","sexual_preference","latitude","longitude","username"],"title":"UserPut"},"UserUpdate":{"properties":{"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"bio":{"anyOf":[{"type":"string","maxLength":255,"minLength":10},{"type":"null"}],"title":"Bio"},"gender":{"anyOf":[{"$ref":"#/components/schemas/Gender"},{"type":"null"}]},"sexual_preference":{"anyOf":[{"$ref":"#/components/schemas/Gender"},{"type":"null"}]},"interests":{"anyOf":[{"items":{"type":"string"},"type":"array","maxItems":5},{"type":"null"}],"title":"Interests"},"images":{"anyOf":[{"items":{"type":"string"},"type":"array","maxItems":5},{"type":"null"}],"title":"Images"},"profile_picture":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Picture"},"latitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Latitude"},"longitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Longitude"},"location_label":{"anyOf":[{"type":"string","maxLength":128},{"type":"null"}],"title":"Location Label"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}